fix: move server index to index file, move routes folder to the top level
