+++ index.js
@@ -1,4 +1,18 @@
-require("./server");
 const { subscribeForLogs } = require("./web3");
 
-subscribeForLogs();
+require("./database");
+const express = require("express");
+const { PORT } = require("./config");
+const configurationRoute = require("./routes/configuration");
+const transactionsRoute = require("./routes/transactions");
+
+const app = express();
+
+app.use(express.json());
+app.use("/configuration", configurationRoute);
+app.use("/transactions", transactionsRoute);
+
+app.listen(PORT, () => {
+  console.log(`Server listening on port ${PORT}`);
+  subscribeForLogs();
+});
